Re-implemented Sided APIs since they had a huge flaw, added more Server orientated Shell commands. Plus bugfixes.

+++ os_files/WolfOS/apis/WAU.moon
@@ -1,6 +1,28 @@
--- WolfOS Account Utilities (SERVER)
+-- WolfOS Account Utilities
 
-WDM = os.getApiSided "WDM"
+WDM = os.getApi "WDM"
+WNC = os.getApi "WNC"
+
+getSide = -> return WDM.readTempData "current_side"
+getModemPort = -> return WDM.readServerData "modem_port"
+getRelayAddress = -> return WDM.readTempData "parent_address"
+getServerState = -> return WDM.readServerData "server_state"
+getServerModuleChannel = ->
+    modules = WDM.readTempData("server_modules") or {}
+    if modules.user
+        return modules.user.channel
+
+getServerAddress = ->
+    channel = getServerModuleChannel!
+    address =  WDM.readServerData "server_address"
+    
+    if address and channel
+        return string.match(address, "(%d+):(%d+)")..":"..channel
+
+thisAddress = ->
+    channel = getServerModuleChannel!
+    if channel
+        return os.getComputerID!..":"..channel
 
 users = {}
 
@@ -18,7 +40,16 @@ generateUID = ->
     return uid
 
 export getUsers = ->
-    users = WDM.readData os.getSystemDir("data").."users.dat" 
+    switch getSide!
+        when "Side.CLIENT"
+            WNC.send getModemPort!, getRelayAddress!, thisAddress!, getServerAddress!, {"data_request", "user_list"}
+            data = WNC.listen getModemPort!, getServerModuleChannel!, 5
+            
+            if data and data[1] == "request_success"
+                users = data[2]
+        when "Side.SERVER"
+            users = WDM.readData os.getSystemDir("data").."users.dat" 
+    
     return users
 
 export exists = (user) ->
@@ -38,44 +69,63 @@ export createUser = (name, hash, type = "user") ->
     if not ok
         error err, 2
     
-    uid = generateUID!
-    table.insert users, {uid: uid, name: name, hash: hash, type: type}
-    WDM.writeData os.getSystemDir("data").."users.dat", users
-    fs.makeDir os.getSystemDir("users")..uid.."/"
+    switch getSide!
+        when "Side.CLIENT"
+            WNC.send getModemPort!, getRelayAddress!, thisAddress!, getServerAddress!, {"data_update", "new_user", name, hash, type}
+        when "Side.SERVER"
+            uid = generateUID!
+            table.insert users, {uid: uid, name: name, hash: hash, type: type}
+            WDM.writeData os.getSystemDir("data").."users.dat", users
+            fs.makeDir os.getSystemDir("users")..uid.."/"
 
 export removeUser = (_user) ->
     ok, err = ftype "string", _user
     if not ok
         error err, 2
     
-    user, i = exists _user
-    if user
-        table.remove users, i
-        WDM.writeData os.getSystemDir("data").."users.dat", users
-        
-        dir = os.getSystemDir("users")..user.uid.."/"
-        if fs.exists dir
-            fs.delete dir
+    switch getSide!
+        when "Side.CLIENT"
+            WNC.send getModemPort!, getRelayAddress!, thisAddress!, getServerAddress!, {"data_update", "remove_user", _user}
+        when "Side.SERVER"
+            user, i = exists _user
+            if user
+                table.remove users, i
+                WDM.writeData os.getSystemDir("data").."users.dat", users
+                dir = os.getSystemDir("users")..user.uid.."/"
+                if fs.exists dir
+                    fs.delete dir
 
 export changeUserData = (_user, k, v) ->
     ok, err = ftype "string, string, string", _user, k, v
     if not ok
         error err, 2
     
-    user, i = exists _user
-    
-    if user
-        if k != "uid"
-            user[k] = v
-            users[i] = user
-            WDM.writeData os.getSystemDir("data").."users.dat", users
+    switch getSide!
+        when "Side.CLIENT"
+            WNC.send getModemPort!, getRelayAddress!, thisAddress!, getServerAddress!, {"data_update", "update_user", _user, k, v}
+        when "Side.SERVER"
+            user, i = exists _user
+            if user
+                if k != "uid"
+                    user[k] = v
+                    users[i] = user
+                    WDM.writeData os.getSystemDir("data").."users.dat", users
 
 export checkLogin = (name, hash) ->
     if ftype "string, string", name, hash
-        user = exists(name) or {}
-        
-        if name == user.name and hash == user.hash
-            WDM.writeTempData true, "local_user"
-            return true, user
+        switch getSide!
+            when "Side.CLIENT"
+                WNC.send getModemPort!, getRelayAddress!, thisAddress!, getServerAddress!, {"login_attempt", name, hash}
+                data = WNC.listen getModemPort!, getServerModuleChannel!, 5
+                
+                if data and data[1] == "login_success"
+                    WDM.writeTempData false, "local_user"
+                    return true, data[2]
+            when "Side.SERVER"
+                user = exists(name) or {}
+                
+                if name == user.name and hash == user.hash
+                    WDM.writeTempData true, "local_user"
+                    return true, user
     
     return false, "Invalid username or password!"
